chore: Move auth store to features/auth
https://harperdb.atlassian.net/browse/STUDIO-588
