+++ src/features/auth/ForgotPassword.tsx
@@ -1,19 +1,19 @@
+import { Button } from '@/components/ui/button';
 import { Form } from '@/components/ui/form/Form';
 import { FormControl } from '@/components/ui/form/FormControl';
 import { FormField } from '@/components/ui/form/FormField';
 import { FormItem } from '@/components/ui/form/FormItem';
 import { FormLabel } from '@/components/ui/form/FormLabel';
 import { FormMessage } from '@/components/ui/form/FormMessage';
+import { Input } from '@/components/ui/input';
 import { zodRequireEmail } from '@/lib/zod/email';
+import { zodResolver } from '@hookform/resolvers/zod';
 import { Link, useNavigate, useSearch } from '@tanstack/react-router';
-import { useForgotPasswordMutation } from '@/features/auth/hooks/useForgotPassword';
 import { useEffect } from 'react';
 import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { z } from 'zod';
-import { Input } from '@/components/ui/input';
-import { Button } from '@/components/ui/button';
 import { toast } from 'sonner';
+import { z } from 'zod';
+import { useForgotPasswordMutation } from './hooks/useForgotPassword';
 
 const ForgotPasswordSchema = z.object({
 	email: zodRequireEmail,
