+++ src/features/organizations/components/NewOrgForm.tsx
@@ -8,10 +8,10 @@ import { FormLabel } from '@/components/ui/form/FormLabel';
 import { FormMessage } from '@/components/ui/form/FormMessage';
 import { Input } from '@/components/ui/input';
 import { currentUserQueryKey } from '@/features/auth/queries/getCurrentUser';
+import { authStore, OverallAppSignIn } from '@/features/auth/store/authStore';
 import { useCreateNewOrganizationMutation } from '@/features/organizations/hooks/useCreateNewOrganization';
 import { NewOrganizationSchema } from '@/features/organizations/mutations/newOrganizationSchema';
 import { useCloudAuth } from '@/hooks/useAuth';
-import { authStore, OverallAppSignIn } from '@/lib/authStore';
 import { collapseKebabsToMaxLength } from '@/lib/string/collapseKebabsToMaxLength';
 import { toKebabCase } from '@/lib/string/to-kebab-case';
 import { zodResolver } from '@hookform/resolvers/zod';
