+++ src/features/profile/index.tsx
@@ -7,10 +7,10 @@ import { FormLabel } from '@/components/ui/form/FormLabel';
 import { FormMessage } from '@/components/ui/form/FormMessage';
 import { Input } from '@/components/ui/input';
 import { logoutOnSuccess } from '@/features/auth/handlers/logoutOnSuccess';
+import { authStore, OverallAppSignIn } from '@/features/auth/store/authStore';
 import { useUpdateUserMutation } from '@/features/profile/mutations/updateUserMutation';
 import { UpdateUserSchema } from '@/features/profile/mutations/updateUserSchema';
 import { useCloudAuth } from '@/hooks/useAuth';
-import { authStore, OverallAppSignIn } from '@/lib/authStore';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useNavigate, useRouter } from '@tanstack/react-router';
 import { Save } from 'lucide-react';
