+++ src/router/rootRoute.ts
@@ -1,7 +1,7 @@
 import { StudioCloud } from '@/components/StudioCloud';
 import { StudioLocal } from '@/components/StudioLocal';
 import { isLocalStudio } from '@/config/constants';
-import { AuthenticatedConnection, EntityIds } from '@/lib/authStore';
+import { AuthenticatedConnection, EntityIds } from '@/features/auth/store/authStore';
 import { QueryClient } from '@tanstack/react-query';
 import { createRootRouteWithContext } from '@tanstack/react-router';
